+++ Assets/Package/Runtime/Assembly.cs
@@ -1,3 +1,4 @@
 using System.Runtime.CompilerServices;
 
 [assembly:InternalsVisibleTo("FinalClick.Services.Editor")]
+[assembly:InternalsVisibleTo("FinalClick.Services.Tests")]